Added loggin functionality and optimized downloading artwork.
Added logging functionality by giving every instance of a Tagger class
a TaggingLogger class with simple to use wrapper functions to write to
formatted log files.

Removed subprocess call to cURL and replaced with Python requests call
to the url which writes all downloaded data to a file named
albumart.jpg in the local directory.